fixed：distributed config  error
