+++ src/main/java/io/jboot/core/http/jboot/JbootHttpImpl.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  * <p>
- *  http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -16,7 +16,6 @@
 package io.jboot.core.http.jboot;
 
 import com.jfinal.log.Log;
-import io.jboot.Jboot;
 import io.jboot.core.http.JbootHttpBase;
 import io.jboot.core.http.JbootHttpRequest;
 import io.jboot.core.http.JbootHttpResponse;
@@ -60,9 +59,7 @@ public class JbootHttpImpl extends JbootHttpBase {
 
 
             if (request.isGetRquest()) {
-                if (Jboot.me().isDevMode()) {
-                    LOG.debug("do get request:" + request.getRequestUrl());
-                }
+
                 connection.setInstanceFollowRedirects(true);
                 connection.connect();
 
@@ -76,10 +73,6 @@ public class JbootHttpImpl extends JbootHttpBase {
              * 处理 post请求
              */
             else if (request.isPostRquest()) {
-
-                if (Jboot.me().isDevMode()) {
-                    LOG.debug("do post request:" + request.getRequestUrl());
-                }
 
                 connection.setRequestMethod("POST");
                 connection.setDoOutput(true);
